% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_ICs.R
\name{EBIC}
\alias{EBIC}
\alias{RBIC}
\alias{RAIC}
\title{Custom IC functions for stepwise models}
\usage{
EBIC(x, ...)

RBIC(x, ...)

RAIC(x, ...)
}
\arguments{
\item{...}{additional args}

\item{fit}{a fitted object}

\item{varnames}{names of variables}

\item{pen_info}{penalty information}

\item{gammafn}{What to use for gamma in formula}

\item{return_df}{should the deg. freedom be returned}
}
\description{
Custom IC functions for stepwise models
}
