% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{cleveland}
\alias{Z}
\alias{S}
\alias{va}
\alias{hungarian}
\alias{irlcs_radon_syn}
\alias{switzerland}
\title{Data sets}
\format{
An object of class \code{data.frame} with 303 rows and 14 columns.

An object of class \code{data.frame} with 294 rows and 14 columns.

An object of class \code{data.frame} with 123 rows and 14 columns.

An object of class \code{data.frame} with 200 rows and 14 columns.

An object of class \code{data.frame} with 1027 rows and 16 columns.

An object of class \code{data.frame} with 442 rows and 6 columns.

An object of class \code{Surv} with 442 rows and 2 columns.
}
\source{
\describe{
  \item{Detrano data}{\url{https://archive.ics.uci.edu/ml/datasets/heart+disease}}
  \item{IRLCS data sets}{\url{https://cheec.uiowa.edu/research/residential-radon-and-lung-cancer-case-control-study}}
  \item{Sheddon}{\url{https://www.gsea-msigdb.org/gsea/msigdb/cards/SHEDDEN_LUNG_CANCER_POOR_SURVIVAL_A6}}
}
}
\usage{
cleveland

hungarian

switzerland

va

irlcs_radon_syn

Z

S
}
\description{
Detrano data sets (cleveland, hungarian, switzerland, va);
The Iowa Radon Lung Cancer Study (irlcs_radon_syn): Data simulated to resemble the IRLCS study;
Sheddon survival data (Z: clinical covariates, S:survival outcome)
}
\references{
\describe{
  \item{Detrano}{
  Detrano, R., Janosi, A., Steinbrunn, W., Pfisterer, M., Schmid, J., Sandhu, S., Guppy, K.,
  Lee, S., & Froelicher, V. (1989). International application of a new probability algorithm
  for the diagnosis of coronary artery disease. American Journal of Cardiology, 64,304--310.
  }
  \item{IRLCS}{
    FIELD, R., SMITH, B., STECK, D. et al. Residential radon exposure and lung cancer:
    Variation in risk estimates using alternative exposure scenarios. J Expo Sci Environ Epidemiol
    12, 197–203 (2002). \url{https://www.nature.com/articles/7500215}
  }
  \item{Shedden}{
  Director's Challenge Consortium for the Molecular Classification of Lung Adenocarcinoma, Shedden, K.,
  Taylor, J. M., Enkemann, S. A., Tsao, M. S., Yeatman, T. J., Gerald, W. L., Eschrich, S., Jurisica, I.,
  Giordano, T. J., Misek, D. E., Chang, A. C., Zhu, C. Q., Strumpf, D., Hanash, S., Shepherd, F. A., Ding,
  K., Seymour, L., Naoki, K., Pennell, N., … Beer, D. G. (2008). Gene expression-based survival prediction
  in lung adenocarcinoma: a multi-site, blinded validation study. Nature medicine, 14(8), 822–827.
  \url{https://www.nature.com/articles/nm.1790}
  }
}
}
\keyword{datasets}
