% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_ICs.R
\name{EBIC}
\alias{EBIC}
\title{Custom IC functions for stepwise models}
\usage{
EBIC(fit, varnames, pen_info, gammafn = NULL, return_df = TRUE, ...)
}
\arguments{
\item{fit}{a fitted object}

\item{varnames}{names of variables}

\item{pen_info}{penalty information}

\item{gammafn}{What to use for gamma in formula}

\item{return_df}{should the deg. freedom be returned}

\item{...}{additional args}
}
\description{
Custom IC functions for stepwise models
}
