% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparseR.R
\name{effect_plot}
\alias{effect_plot}
\alias{effect_plot.sparseR}
\alias{effect_plot.sparseRBIC}
\title{Plot relevant effects of a sparseR object}
\usage{
effect_plot(fit, ...)

\method{effect_plot}{sparseR}(
  fit,
  coef_name,
  at = c("cvmin", "cv1se"),
  by = NULL,
  by_levels,
  nn = 101,
  plot.args = list(),
  resids = TRUE,
  legend_location = "bottomright",
  ...
)

\method{effect_plot}{sparseRBIC}(
  fit,
  coef_name,
  by = NULL,
  by_levels,
  nn = 101,
  plot.args = list(),
  resids = TRUE,
  legend_location = "bottomright",
  ...
)
}
\arguments{
\item{fit}{a `sparseR` object}

\item{...}{additional arguments}

\item{coef_name}{The name of the coefficient to plot along the x-axis}

\item{at}{value of lambda to use}

\item{by}{the variable(s) involved in the (possible) interaction}

\item{by_levels}{values to cut continuous by variable (defaults to 3 quantiles)}

\item{nn}{number of points to plot along prediction line}

\item{plot.args}{list of arguments passed to the plot itself}

\item{resids}{should residuals be plotted or not?}

\item{legend_location}{location for legend passed to `legend`}
}
\value{
nothing returned

Nothing (invisible) returned
}
\description{
Plot relevant effects of a sparseR object
}
