% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparseR.R
\name{plot.sparseR}
\alias{plot.sparseR}
\title{Plot relevant properties of sparseR objects}
\usage{
\method{plot}{sparseR}(x, plot_type = c("both", "cv", "path"), cols = NULL, log.l = TRUE, ...)
}
\arguments{
\item{x}{a `sparseR` object}

\item{plot_type}{should the solution path, CV results, or both be plotted?}

\item{cols}{option to specify color of groups}

\item{log.l}{should the x-axis (lambda) be logged?}

\item{...}{extra plotting options}
}
\value{
nothing returned
}
\description{
Plot relevant properties of sparseR objects
}
