% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseR.R
\name{predict.sparseR}
\alias{predict.sparseR}
\alias{coef.sparseR}
\title{Predict coefficients or responses for sparseR object}
\usage{
\method{predict}{sparseR}(object, newdata, lambda, at = c("cvmin", "cv1se"), ...)

\method{coef}{sparseR}(object, lambda, at = c("cvmin", "cv1se"), ...)
}
\arguments{
\item{object}{sparseR object}

\item{newdata}{new data on which to make predictions}

\item{lambda}{a particular value of lambda to predict with}

\item{at}{a "smart" guess to use for lambda}

\item{...}{additional arguments passed to predict.ncvreg}
}
\value{
predicted outcomes for `newdata` (or coefficients)
  at specified (or smart) lambda value
}
\description{
Predict coefficients or responses for sparseR object
}
