% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseR.R
\name{print.sparseR}
\alias{print.sparseR}
\title{Print sparseR object}
\usage{
\method{print}{sparseR}(x, prep = FALSE, ...)
}
\arguments{
\item{x}{a sparseR object}

\item{prep}{Should the SR set-up information be printed as well?}

\item{...}{additional arguments passed to print.ncvreg}
}
\value{
returns x invisibly
}
\description{
Print sparseR object
}
