% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseR-package.R
\docType{package}
\name{sparseR-package}
\alias{sparseR-package}
\alias{_PACKAGE}
\title{sparseR: Implement ranked sparsity for selecting interactions and polynomials}
\description{
The \code{sparseR} package implements various techniques for selecting from
a set of interaction and polynomial terms under ranked sparsity. Additional
tools for data pre-processing, post-selection inference, and visualization
are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://petersonr.github.io/sparseR/}
  \item \url{https://github.com/petersonR/sparseR/}
}

}
\author{
\strong{Maintainer}: Ryan Andrew Peterson \email{ryan.a.peterson@cuanschutz.edu} (\href{https://orcid.org/0000-0002-4650-5798}{ORCID})

}
