% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseRBIC_sampsplit.R
\name{sparseRBIC_sampsplit}
\alias{sparseRBIC_sampsplit}
\title{Sample split procedure for stepwise regression}
\usage{
sparseRBIC_sampsplit(srbic_fit, S = 100, quiet = FALSE)
}
\arguments{
\item{srbic_fit}{An object fitted by sparseRBIC_step}

\item{S}{Number of splitting iterations}

\item{quiet}{Should the display of a progress bar be silenced?}
}
\value{
a list containing:

\item{results}{a tibble containing coefficients, p-values, selection pct}
\item{splits}{a tibble of different split-based coefficients}
}
\description{
Runs multiple on models selection procedures using RBIC to
achieve valid inferential results post-selection
}
