% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseR.R
\name{summary.sparseR}
\alias{summary.sparseR}
\title{Summary of sparseR model coefficients}
\usage{
\method{summary}{sparseR}(object, lambda, at = c("cvmin", "cv1se"), ...)
}
\arguments{
\item{object}{a sparseR object}

\item{lambda}{a particular value of lambda to predict with}

\item{at}{a "smart" guess to use for lambda}

\item{...}{additional arguments to be passed to summary.ncvreg}
}
\value{
an object of class `summary.ncvreg` at specified or smart value of
  lambda.
}
\description{
Summary of sparseR model coefficients
}
