% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_ICs.R
\name{EBIC}
\alias{EBIC}
\alias{RBIC}
\alias{RAIC}
\alias{EBIC.default}
\alias{RBIC.default}
\alias{RAIC.default}
\title{Custom IC functions for stepwise models}
\usage{
EBIC(...)

\method{EBIC}{default}(fit, varnames, pen_info, gammafn = NULL, return_df = TRUE, ...)

RBIC(fit, ...)

\method{RBIC}{default}(fit, varnames, pen_info, gammafn = NULL, return_df = TRUE, ...)

RAIC(fit, ...)

\method{RAIC}{default}(fit, varnames, pen_info, gammafn = NULL, return_df = TRUE, ...)
}
\arguments{
\item{...}{additional args}

\item{fit}{a fitted object}

\item{varnames}{names of variables}

\item{pen_info}{penalty information}

\item{gammafn}{What to use for gamma in formula}

\item{return_df}{should the deg. freedom be returned}
}
\value{
A vector of values for the criterion requested, and the degrees of
  freedom (appended to front of vector) if return_df == TRUE.
}
\description{
Custom IC functions for stepwise models
}
