% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseRBIC_bootstrap.R
\name{sparseRBIC_bootstrap}
\alias{sparseRBIC_bootstrap}
\title{Bootstrap procedure for stepwise regression}
\usage{
sparseRBIC_bootstrap(srbic_fit, B = 100, quiet = FALSE)
}
\arguments{
\item{srbic_fit}{An object fitted by sparseRBIC_step}

\item{B}{Number of bootstrap samples}

\item{quiet}{Should the display of a progress bar be silenced?}
}
\value{
a list containing:

\item{results}{a tibble containing coefficients, p-values, selection pct}
\item{bootstraps}{a tibble of bootstrapped coefficients}
}
\description{
Runs bootstrap on models selection procedure using RBIC to
find bootstrapped standard error (smoothed, see Efron 2014)
as well as selection percentage across candidate variables.
(experimental)
}
